ReviewMovie.xibを使い回すため、管理するクラスを生成
