+++ MovieReviewMVP/Cells/Xib/ReviewMovie.xib
@@ -8,7 +8,7 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="ReviewMovieViewController" customModule="MovieReviewMVP" customModuleProvider="target">
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="ReviewMovieOwner" customModule="MovieReviewMVP" customModuleProvider="target">
             <connections>
                 <outlet property="backgroundImageView" destination="TUf-tF-Y2T" id="8sM-Bz-h6Y"/>
                 <outlet property="movieImageView" destination="rwh-PO-286" id="EqQ-VA-MMT"/>
